+++ docs/package.go
@@ -178,6 +178,10 @@ func NewProjectArticle() *Element {
 		ExampleSequenceDiagram().Inline(),
 		LoadFile("sequence_example.go", 1, -1),
 
+		H3("C4 diagrams"),
+		ExampleC4SystemContextDiagram().Inline(),
+		LoadFile("c4_example.go", 1, -1),
+		
 		H3("Gantt chart"),
 		ExampleGanttChart().Inline(),