+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.34.0-dev]
+
+- Fix shape.Container width, wrap to grouped objects
+- Add C4 example
+
 ## [0.33.0] 2023-10-21
 
 - Use gregoryv/web v0.25.0